replaced uint with int - less casting

+++ ClientDecisionService/VW/VWJson.cs
@@ -11,7 +11,7 @@ using VW.Serializer;
 namespace Microsoft.Research.MultiWorldTesting.ClientLibrary
 {
     internal class VWJsonPolicy : 
-        VWBaseContextMapper<VowpalWabbitThreadedPrediction, VowpalWabbit, string, uint>, 
+        VWBaseContextMapper<VowpalWabbitThreadedPrediction, VowpalWabbit, string, int>, 
         IPolicy<string>
     {
         internal VWJsonPolicy(Stream vwModelStream = null)
@@ -19,12 +19,12 @@ namespace Microsoft.Research.MultiWorldTesting.ClientLibrary
         {
         }
 
-        protected override Decision<uint> MapContext(VowpalWabbit vw, string context)
+        protected override Decision<int> MapContext(VowpalWabbit vw, string context)
         {
             using (var vwJson = new VowpalWabbitJsonSerializer(vw))
             using (VowpalWabbitExampleCollection vwExample = vwJson.ParseAndCreate(context))
             {
-                var action = (uint)vwExample.Predict(VowpalWabbitPredictionType.CostSensitive);
+                var action = (int)vwExample.Predict(VowpalWabbitPredictionType.CostSensitive);
                 var state = new VWState { ModelId = vw.ID };
 
                 return Decision.Create(action, state);
@@ -33,7 +33,7 @@ namespace Microsoft.Research.MultiWorldTesting.ClientLibrary
     }
 
     internal class VWJsonRanker : 
-        VWBaseContextMapper<VowpalWabbitThreadedPrediction, VowpalWabbit, string, uint[]>, 
+        VWBaseContextMapper<VowpalWabbitThreadedPrediction, VowpalWabbit, string, int[]>, 
         IRanker<string>, INumberOfActionsProvider<string>
     {
         internal VWJsonRanker(Stream vwModelStream = null)
@@ -41,7 +41,7 @@ namespace Microsoft.Research.MultiWorldTesting.ClientLibrary
         {
         }
 
-        protected override Decision<uint[]> MapContext(VowpalWabbit vw, string context)
+        protected override Decision<int[]> MapContext(VowpalWabbit vw, string context)
         {
             using (var vwJson = new VowpalWabbitJsonSerializer(vw))
             using (VowpalWabbitExampleCollection vwExample = vwJson.ParseAndCreate(context))
@@ -49,7 +49,7 @@ namespace Microsoft.Research.MultiWorldTesting.ClientLibrary
                 int[] vwMultilabelPredictions = vwExample.Predict(VowpalWabbitPredictionType.Multilabel);
 
                 // VW multi-label predictions are 0-based
-                var actions = vwMultilabelPredictions.Select(a => (uint)(a + 1)).ToArray();
+                var actions = vwMultilabelPredictions.Select(a => a + 1).ToArray();
                 var state = new VWState { ModelId = vw.ID };
 
                 return Decision.Create(actions, state);
