+++ explore/GenericExplorer.cs
@@ -19,27 +19,27 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
 	/// distribution over actions desired, and it will draw from that.
 	/// </remarks>
 	/// <typeparam name="TContext">The Context type.</typeparam>
-    public class GenericExplorer<TContext> : BaseExplorer<TContext, uint, GenericExplorerState, float[]>
+    public class GenericExplorer<TContext> : BaseExplorer<TContext, int, GenericExplorerState, float[]>
 	{
 		/// <summary>
 		/// The constructor is the only public member, because this should be used with the MwtExplorer.
 		/// </summary>
 		/// <param name="defaultScorer">A function which outputs the probability of each action.</param>
 		/// <param name="numActions">The number of actions to randomize over.</param>
-        public GenericExplorer(IContextMapper<TContext, float[]> defaultScorer, uint numActions = uint.MaxValue)
+        public GenericExplorer(IContextMapper<TContext, float[]> defaultScorer, int numActions = int.MaxValue)
             : base(defaultScorer, numActions)
 		{
         }
 
-        public override Decision<uint, GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
+        public override Decision<int, GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
         {
             var random = new PRG(saltedSeed);
 
             // Invoke the default scorer function
             Decision<float[]> policyDecision = this.contextMapper.MapContext(context);
             if (policyDecision == null)
-                return Decision.Create<uint, GenericExplorerState, float[]>(
-                    (uint)random.UniformInt(1, this.numActionsFixed),
+                return Decision.Create<int, GenericExplorerState, float[]>(
+                    random.UniformInt(1, this.numActionsFixed),
                     new GenericExplorerState
                     {
                         Probability = 1f
@@ -49,8 +49,8 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
 
             float[] weights = policyDecision.Value;
 
-            uint numWeights = (uint)weights.Length;
-            if (this.numActionsFixed != uint.MaxValue && numWeights != this.numActionsFixed)
+            int numWeights = weights.Length;
+            if (this.numActionsFixed != int.MaxValue && numWeights != this.numActionsFixed)
             {
                 throw new ArgumentException("The number of weights returned by the scorer must equal number of actions");
             }
@@ -75,14 +75,14 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
 
             float sum = 0f;
             float actionProbability = 0f;
-            uint actionIndex = numWeights - 1;
+            int actionIndex = numWeights - 1;
             for (int i = 0; i < numWeights; i++)
             {
                 weights[i] = weights[i] / total;
                 sum += weights[i];
                 if (sum > draw)
                 {
-                    actionIndex = (uint)i;
+                    actionIndex = (int)i;
                     actionProbability = weights[i];
                     break;
                 }
@@ -108,7 +108,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
     /// </remarks>
     /// <typeparam name="TContext">The Context type.</typeparam>
     public sealed class GenericExplorerSampleWithoutReplacement<TContext> 
-        : BaseExplorer<TContext, uint[], GenericExplorerState, float[]>
+        : BaseExplorer<TContext, int[], GenericExplorerState, float[]>
     {
         private readonly GenericExplorer<TContext> explorer;
 
@@ -117,7 +117,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
         /// </summary>
         /// <param name="defaultScorer">A function which outputs the probability of each action.</param>
         /// <param name="numActions">The number of actions to randomize over.</param>
-        public GenericExplorerSampleWithoutReplacement(IContextMapper<TContext, float[]> defaultScorer, uint numActions = uint.MaxValue)
+        public GenericExplorerSampleWithoutReplacement(IContextMapper<TContext, float[]> defaultScorer, int numActions = int.MaxValue)
              : base(defaultScorer, numActions)
         {
             this.explorer = new GenericExplorer<TContext>(defaultScorer, numActions);
@@ -129,7 +129,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
             this.explorer.EnableExplore(explore);
         }
 
-        public override Decision<uint[], GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
+        public override Decision<int[], GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
         {
             var random = new PRG(saltedSeed);
 
@@ -139,7 +139,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
             float[] weights = decision.MapperDecision.Value;
 
             float actionProbability = 0f;
-            uint[] chosenActions = MultiActionHelper.SampleWithoutReplacement(weights, (uint)weights.Length, random, ref actionProbability);
+            int[] chosenActions = MultiActionHelper.SampleWithoutReplacement(weights, weights.Length, random, ref actionProbability);
 
             // action id is one-based
             return Decision.Create(chosenActions,
