+++ explore/SoftmaxExplorer.cs
@@ -13,7 +13,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
 	/// you to do that.
 	/// </remarks>
 	/// <typeparam name="TContext">The Context type.</typeparam>
-    public sealed class SoftmaxExplorer<TContext> : BaseExplorer<TContext, uint, GenericExplorerState, float[]>
+    public sealed class SoftmaxExplorer<TContext> : BaseExplorer<TContext, int, GenericExplorerState, float[]>
 	{
 	    private readonly float lambda;
 
@@ -23,21 +23,21 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
 		/// <param name="defaultScorer">A function which outputs a score for each action.</param>
 		/// <param name="lambda">lambda = 0 implies uniform distribution. Large lambda is equivalent to a max.</param>
 		/// <param name="numActions">The number of actions to randomize over.</param>
-        public SoftmaxExplorer(IContextMapper<TContext, float[]> defaultScorer, float lambda, uint numActions)
+        public SoftmaxExplorer(IContextMapper<TContext, float[]> defaultScorer, float lambda, int numActions)
             : base(defaultScorer, numActions)
         {
             this.lambda = lambda;
         }
 
-        public override Decision<uint, GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
+        public override Decision<int, GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
         {
             var random = new PRG(saltedSeed);
 
             // Invoke the default scorer function
             Decision<float[]> policyDecision = this.contextMapper.MapContext(context);
             if (policyDecision == null)
-                return Decision.Create<uint, GenericExplorerState, float[]>(
-                    (uint)random.UniformInt(1, this.numActionsFixed),
+                return Decision.Create<int, GenericExplorerState, float[]>(
+                    random.UniformInt(1, this.numActionsFixed),
                     new GenericExplorerState
                     {
                         Probability = 1f
@@ -46,8 +46,8 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
                     shouldRecord: true);
 
             float[] scores = policyDecision.Value;
-            uint numScores = (uint)scores.Length;
-            if (this.numActionsFixed != uint.MaxValue && numScores != this.numActionsFixed)
+            int numScores = scores.Length;
+            if (this.numActionsFixed != int.MaxValue && numScores != this.numActionsFixed)
             {
                 throw new ArgumentException("The number of scores returned by the scorer must equal number of actions");
             }
@@ -57,7 +57,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
             float maxScore = scores.Max();
 
             float actionProbability = 0f;
-            uint actionIndex = 0;
+            int actionIndex = 0;
             if (this.explore)
             {
                 // Create a normalized exponential distribution based on the returned scores
@@ -81,7 +81,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
                     sum += scores[i];
                     if (sum > draw)
                     {
-                        actionIndex = (uint)i;
+                        actionIndex = i;
                         actionProbability = scores[i];
                         break;
                     }
@@ -95,7 +95,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
                     if (maxScore < scores[i])
                     {
                         maxScore = scores[i];
-                        actionIndex = (uint)i;
+                        actionIndex = i;
                     }
                 }
                 actionProbability = 1f; // Set to 1 since we always pick the highest one.
@@ -112,7 +112,7 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
     }
 
     public sealed class SoftmaxSampleWithoutReplacementExplorer<TContext>
-        : BaseExplorer<TContext, uint[], GenericExplorerState, float[]>
+        : BaseExplorer<TContext, int[], GenericExplorerState, float[]>
     {
         private readonly SoftmaxExplorer<TContext> explorer;
 
@@ -123,9 +123,9 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
 		/// <param name="lambda">lambda = 0 implies uniform distribution. Large lambda is equivalent to a max.</param>
 		/// <param name="numActions">The number of actions to randomize over.</param>
         public SoftmaxSampleWithoutReplacementExplorer(IContextMapper<TContext, float[]> defaultScorer, float lambda)
-            : base(defaultScorer, uint.MaxValue)
+            : base(defaultScorer, int.MaxValue)
         {
-            this.explorer = new SoftmaxExplorer<TContext>(defaultScorer, lambda, uint.MaxValue);
+            this.explorer = new SoftmaxExplorer<TContext>(defaultScorer, lambda, int.MaxValue);
         }
 
         public override void EnableExplore(bool explore)
@@ -134,18 +134,18 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
             this.explorer.EnableExplore(explore);
         }
 
-        public override Decision<uint[], GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
+        public override Decision<int[], GenericExplorerState, float[]> MapContext(ulong saltedSeed, TContext context)
         {
             var decision = this.explorer.MapContext(saltedSeed, context);
             var scores = decision.MapperDecision.Value;
-            uint numActionsVariable = (uint)scores.Length;
+            int numActionsVariable = scores.Length;
 
             if (scores == null || scores.Length < 1)
             {
                 throw new ArgumentException("Scores returned by default policy must not be empty.");
             }
 
-            uint[] chosenActions;
+            int[] chosenActions;
             float actionProbability = decision.ExplorerState.Probability;
 
             if (this.explore)
@@ -156,14 +156,14 @@ namespace Microsoft.Research.MultiWorldTesting.ExploreLibrary
             else
             {
                 // avoid linq to optimize perf
-                chosenActions = new uint[numActionsVariable];
+                chosenActions = new int[numActionsVariable];
                 for (int i = 1; i <= numActionsVariable; i++)
                 {
-                    chosenActions[i] = (uint)i;
+                    chosenActions[i] = i;
                 }
 
                 // swap max-score action with the first one
-                uint firstAction = chosenActions[0];
+                int firstAction = chosenActions[0];
                 chosenActions[0] = chosenActions[decision.Value];
                 chosenActions[decision.Value] = firstAction;
             }
